+++ sparse_ho/__init__.py
@@ -1,5 +1,5 @@
 from .ho import grad_search
-from .ho import grad_search_wolfe
+# from .ho import grad_search_wolfe
 from .backward import Backward
 from .forward import Forward
 from .implicit_forward import ImplicitForward
